refactor(py/executor): extract fanout and find_actions from command.py
